changes done

+++ Tips.Grin.Api/Entities/DTOs/OpenGrinPartsDto.cs
@@ -7,7 +7,7 @@ namespace Tips.Grin.Api.Entities.DTOs
         public int Id { get; set; }
         public string? ItemNumber { get; set; }
         public string? Description { get; set; }
-        public decimal ItemType { get; set; }
+        public int ItemType { get; set; }
         public string? UOM { get; set; }
         [Precision(18, 3)]
         public decimal Qty { get; set; }
@@ -21,7 +21,7 @@ namespace Tips.Grin.Api.Entities.DTOs
         {
             public string? ItemNumber { get; set; }
             public string? Description { get; set; }
-            public decimal ItemType { get; set; }
+            public int ItemType { get; set; }
             public string? UOM { get; set; }
             [Precision(18, 3)]
             public decimal Qty { get; set; }
@@ -36,7 +36,7 @@ namespace Tips.Grin.Api.Entities.DTOs
             public int Id { get; set; }
             public string? ItemNumber { get; set; }
             public string? Description { get; set; }
-            public decimal ItemType { get; set; }
+            public int ItemType { get; set; }
             public string? UOM { get; set; }
             [Precision(18, 3)]
             public decimal Qty { get; set; }
